Patch small issue in dbPrepareStatements
This method assumes that you give it either precompiled statements or
strings. This adds a bit of a guard that ensures that you don't pass
any scalars through.

At the same time, a small change is being made to the dbRawQuery()
function so that it uses a sensible default if no action is provided in
the call, which makes it easier to quickly use it without getting weird
looking logs.