Additions to the predefined hooks for aegis
This is mostly a structural change in the file but it does add in some
extra assertions taht will come in handy later.

Parts of the code were refactored to be function calls so that they can
be easily reused in other package test suites. At a later date this
will be wrapped into a library of its own.