Include unit tests for statement execution
This includes some but not all tests related to doing raw fetch queries
to validate that they all work as expected.

Tests related to validating that the SQL is valid are not currently
implemented because our simplistic scheme for checking if a function
throws only works for a sync function, and there is imminently going to
be a new version of the test library that supports it directly.