Update workflow to fix version numbers
Something that went initially unnoticed is that the tag that we're
pushing starts with a "v" and so we do not want to populate it directly
into the package.json file.

Although this seems to work fine, it is not technically in spec as far
as I can tell.

As such, this change attempts to resolve that by stripping the v off
of it and storing it in a separate variable, which the internet tells
me is entirely viable and not a bad idea at all.