Update our unit tests to use the new Aegis lib
@axel669 has released a newer version of Aegis that reworks how the
test framework functions.

The new changes allow us to control the order that test files are run,
we can label individual tests as desired, and we can stop the test run
as soon as their is a failure in a few cases.

Here we're stopping on errors at the collection level so that we can
first test core functionality and then functionality that uses the core,
with the knowledge that if the core does not work, no other tests are
going to pass anyway.

Note that as of the current time, Aegis 0.3.0 has a couple of minor bugs
in it that will stop the tests from running in a pipeline, so a future
commit will cycle back and update the library version.