Implement new functionality for named binds
Using the code added in the prior commits, tie the new parsing logic
into the existing API. This is fairly straight forward since all of
the work happens elsewhere for the most part, but this does require us
to alter our tests and include a new set that will allow us to test that
all of the parts work.

Arguably the metadata parsing test should have been included in one of
the prior commits, but until all of the other tests also worked I did
not fully trust that everything was kosher.

More tests should be included since there are still some things that are
not properly exercised, but this is good enough for now.