Implement custom error classes
We will be embarking upon an epic journey by which we allow the library
to use named binds, to get around an issue that we have discovered in
which our node-sql-parser that we use for the rollup plugin not having
support for ?# style binds.

To begin this journey, we will include some error classes that we can
use so that when we trigger errors related to invalid SQL or invalid
attempts to bind, we have custom exceptions we can throw for that.