Allow binding from the rollup plugin
The rollup plugin has been augmented so that it exports a few functions,
and so that the main one is more useful.

The default export (also exported under the name "statements") is a
function that can optionally take bind arguments and will attempt to
bind the statements that it returns back.

This works in two ways:

1. If there is a single bindable statement, any number of binds can be
   passed, and the result will be having that statement bound multiple
   times.

3. If there is more than one bindable statement, then you must pass the
   same number of binds as the number of statements that need them, and
   they will be bound in order.

The return value of the function is either a single statement or an
array of statements; that is, if the file contains a single statement,
then the result is always a single statement and not an array with one
element.

In addition, each file also provided a fetch(), fetchOne() and execute
() function which behaves as the above, but takes an extra argument
that is the "action". These functions behave the same as the API
functions of similar names, and are a convenience around being able to
bind and execute in a single step.