Allow binding from the rollup plugin
The rollup plugin has been augmented so that the function that it
exports can optionally take bind arguments and will attempt to bind the
statements that it returns back.

This works in two ways:

1. If there is a single bindable statement, any number of binds can be
   passed, and the result will be having that statement bound multiple
   times.

3. If there is more than one bindable statement, then you must pass the
   same number of binds as the number of statements that need them, and
   they will be bound in order.

In addition, the plugin is now a bit more intelligent in that it only
spits out one copy of the setup code no matter how many SQL files are
imported.

More tests are forthcoming on this to ensure that the plugin works; we
still need more tests for binding in various circumstances, and also
for testing execution.