Inline our SQL parser library
We are about to embark on some new features in the library, to include
support for named bind parameters. In preparing for this change it has
been discovered that in fact the SQLite parser in the base library does
not properly handle numbered bind placeholders, and it also does not
handle binds in the LIMIT and OFFSET.

The easiest course of action to fix this in the near term (and also it
makes us less dependent on other things), we have forked the
node-sql-parser library, made the change, and included our own custom
build that gives us an ESM version of just the SQLite portion of the
code, which is now inlined here.