Export the SQLite parser instance
For use in other packages, particularly a testing library that we are
working on, export the internal Parser object for SQLite that is
inferred from the node-sql-parser package (see the README).

This is also potentially generally useful if something needs to compile
SQLite code directly into an AST for any reason.