Allow statements to know if they produce results
This change augments the statement processor so that for any given
statement that gets processed it flags to indicate if this statement
could possibly return a result.

The distinction here is statements which COULD (but may not) return a
result, such as SELECT, and statements that CANNOT return a result,
like INSERT.

Since we support SQLite solely, and SQLite supports adding RETURNING to
some statements to indicate that it should return back the affected
results, the AST is also analyzed to mark such statements as
potentially returning a result, like an INSERT RETURNING and similar.