Include a new API for working with batches
Since the case comes up from time to time where a batch of statements
exists in which some inserts, updates, deletes, etc are in the batch
along with something that can actually return a value, this change
introduces a new dbFetchFirst() API which behavces identically to
dbFetchOne() when the input is a single statement but behaves
differently for a batch.

When the statement is a batch, this API uses the functionality from the
previous commit that marks statements as knowing if they can produce a
result or not to select from the list of input statements the first one
that contains potential usable results, and then behaves as if that was
the only statement.

This is implemented in the rollup plugin as well. At the same time, a
tweak was made to the rollup plugin that makes the usage of the internal
functions map better, since the call signature is meant to assume that
an array of functions is a batch, even if it's a single statement.

This does not change anything, but it makes the intent clearer.